+++ platform/src/sdk/models/shared/clientratelimitresetwebhook.ts
@@ -37,7 +37,7 @@ export class ClientRateLimitResetWebhook extends SpeakeasyBase {
     data?: ClientRateLimitResetWebhookData;
 
     /**
-     * A human readable message about the webhook.
+     * A human-readable message about the webhook.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "Message" })