+++ platform/src/sdk/models/shared/pushoperationtimedoutwebhook.ts
@@ -51,7 +51,7 @@ export class PushOperationTimedOutWebhook extends SpeakeasyBase {
     dataConnectionId?: string;
 
     /**
-     * A human readable message about the webhook.
+     * A human-readable message about the webhook.
      */
     @SpeakeasyMetadata()
     @Expose({ name: "Message" })