+++ platform/src/sdk/sdk.ts
@@ -63,9 +63,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "2.3.0";
-    genVersion = "2.253.0";
-    userAgent = "speakeasy-sdk/typescript 2.3.0 2.253.0 3.0.0 @codat/platform";
+    sdkVersion = "2.4.0";
+    genVersion = "2.277.0";
+    userAgent = "speakeasy-sdk/typescript 2.4.0 2.277.0 3.0.0 @codat/platform";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);