+++ platform/RELEASES.md
@@ -89,3 +89,13 @@ Based on:
 - [typescript v2.3.0] platform
 ### Releases
 - [NPM v2.3.0] https://www.npmjs.com/package/@codat/platform/v/2.3.0 - platform
+
+## 2024-03-06 11:04:26
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.201.1 (2.277.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.4.0] platform
+### Releases
+- [NPM v2.4.0] https://www.npmjs.com/package/@codat/platform/v/2.4.0 - platform